requirements.txt
python-telegram-bot[ext]>=21.3
web3>=6.15.0
aiohttp>=3.9.0
python-dotenv>=1.0.0
cachetools>=5.3.0
APScheduler>=3.10.4
